Add GCS to factory method, and refactor GCE for consistant project structure
