+++ pkg/clients/BUILD.bazel
@@ -1,26 +1,13 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "clients.go",
-        "gce.go",
-    ],
+    srcs = ["clients.go"],
     importpath = "github.com/googleinterns/cloudai-gcp-test-resource-reaper/pkg/clients",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/resources:go_default_library",
-        "//proto:go_default_library",
-        "@org_golang_google_api//compute/v1:go_default_library",
-        "@org_golang_google_api//option:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["gce_test.go"],
-    embed = [":go_default_library"],
-    deps = [
+        "//pkg/clients/gce:go_default_library",
+        "//pkg/clients/gcs:go_default_library",
         "//pkg/resources:go_default_library",
         "//proto:go_default_library",
         "@org_golang_google_api//option:go_default_library",
