+++ pkg/clients/gce/BUILD.bazel
@@ -2,24 +2,24 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["gcs_client.go"],
-    importpath = "github.com/googleinterns/cloudai-gcp-test-resource-reaper/pkg/clients/gcs",
+    srcs = ["gce_client.go"],
+    importpath = "github.com/googleinterns/cloudai-gcp-test-resource-reaper/pkg/clients/gce",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/resources:go_default_library",
         "//proto:go_default_library",
-        "@com_google_cloud_go_storage//:go_default_library",
+        "@org_golang_google_api//compute/v1:go_default_library",
         "@org_golang_google_api//option:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["gcs_test.go"],
+    srcs = ["gce_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//pkg/resources:go_default_library",
-        "//pkg/utils:go_default_library",
         "//proto:go_default_library",
+        "@org_golang_google_api//option:go_default_library",
     ],
 )
