+++ pkg/clients/gce/gce_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package clients
+package gce
 
 import (
 	"context"
